ajustes no mapeamento das classes

+++ src/main/java/com/camilasoares/cursomc/domain/enums/Perfil.java
@@ -34,7 +34,7 @@ public enum Perfil {
 		this.descricao = descricao;
 	}
 	
-	public static Perfil toEnum(Long cod){
+	public static Perfil toEnum(Integer cod){
 		if(cod == null){
 			return null;
 		}
