+++ src/main/java/com/camilasoares/cursomc/repositories/PedidoRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.camilasoares.cursomc.domain.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
+
 
 }
